+++ belpy.py
@@ -48,12 +48,9 @@ from lib.sfm.absolute_orientation import (
     Space_resection,
 )
 from lib.read_config import parse_yaml_cfg
-from lib.point_clouds import (
-    create_point_cloud,
-    write_ply,
-)
 from lib.utils.utils import (
     AverageTimer,
+    homography_warping,
     build_dsm,
     generate_ortophoto,
 )
@@ -61,7 +58,9 @@ from lib.utils.inizialize_variables import Inizialization
 from lib.visualization import (
     display_point_cloud,
     make_focal_length_variation_plot,
+    make_camera_angles_plot,
     plot_features,
+    imshow_cv,
 )
 from lib.import_export.export2bundler import write_bundler_out
 from lib.metashape.metashape import (
@@ -71,44 +70,7 @@ from lib.metashape.metashape import (
 )
 
 
-from lib.utils.utils import PrintMatrix
-from lib.visualization import imshow_cv
-
 logger = logging.getLogger("Belpy")
-
-
-def homography_warping(
-    cam_0: np.ndarray,
-    cam_1: np.ndarray,
-    image: np.ndarray,
-    out_path: str = None,
-) -> np.ndarray:
-
-    print("performing homography warping based on extrinsics matrix...")
-
-    # Create deepcopies to not modify original data
-    cam_0_ = deepcopy(cam_0)
-    cam_1_ = deepcopy(cam_1)
-
-    T = np.linalg.inv(cam_0_.pose)
-    cam_0_.update_extrinsics(cam_0_.pose_to_extrinsics(T @ cam_0_.pose))
-    cam_1_.update_extrinsics(cam_1_.pose_to_extrinsics(T @ cam_1_.pose))
-
-    R = cam_1_.R
-    K = cam_1_.K
-    H = (K @ R) @ np.linalg.inv(K)
-
-    image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
-    w, h = image.shape[:2]
-    warped_image = cv2.warpPerspective(image, H, (h, w))
-    if out_path is not None:
-        cv2.imwrite(out_path, warped_image)
-        print(f"Warped image {out_path.stem} exported correctely")
-    else:
-        imshow_cv(warped_image, convert_RGB2BRG=False)
-
-    return cv2.cvtColor(warped_image, cv2.COLOR_BGR2RGB)
-
 
 timer_global = AverageTimer(newline=True)
 
@@ -155,25 +117,30 @@ for epoch in cfg.proc.epoch_to_process:
             fname = list(path.glob("*.pickle"))
             if len(fname) < 1:
                 raise FileNotFoundError(
-                    f"No pickle file present in the epoch directory {epochdir}"
+                    f"No pickle file found in the epoch directory {epochdir}"
                 )
             if len(fname) > 1:
                 raise FileNotFoundError(
-                    f"More than one pickle file present in the epoch directory {epochdir}"
+                    f"More than one pickle file is present in the epoch directory {epochdir}"
                 )
 
             with open(fname[0], "rb") as f:
-                loaded_features = pickle.load(f)
+                try:
+                    loaded_features = pickle.load(f)
+                except:
+                    raise FileNotFoundError(
+                        f"Invalid pickle file in epoch directory {epochdir}"
+                    )
 
             for cam, feats in loaded_features.items():
-                features[cam].append(feats)
+                features[cam].insert(epoch, feats)
 
             # with open(cfg.paths.last_match_path, "rb") as f:
             #     features = pickle.load(f)
             #     print("Loaded previous matches")
 
         except FileNotFoundError as err:
-            logger.exception(err.args[0])
+            logger.exception(err)
 
             print("Performing new matching and tracking...")
             features = MatchingAndTracking(
@@ -345,7 +312,7 @@ for epoch in cfg.proc.epoch_to_process:
         cam = "p2"
         image = images[cam][epoch]
         out_path = f"res/warped/{images[cam].get_image_name(epoch)}"
-        homography_warping(cameras[cam][0], cameras[cam][epoch], image, out_path)
+        homography_warping(cameras[cam][0], cameras[cam][epoch], image, out_path, timer)
 
     timer.print(f"Epoch {epoch} completed")
 
@@ -362,6 +329,7 @@ if cfg.other.do_viz:
 
     # Display estimated focal length variation
     make_focal_length_variation_plot(focals, "res/focal_lenghts.png")
+    make_camera_angles_plot(cameras, "res/angles.png")
 
 
 # Debug
@@ -375,43 +343,6 @@ if cfg.other.do_viz:
 
 
 # Camera angles plot
-
-
-def make_camera_angles_plot(cameras):
-    import matplotlib.pyplot as plt
-    import matplotlib
-
-    matplotlib.use("TkAgg")
-
-    omega, phi, kappa = {}, {}, {}
-    for key, cam_list in cameras.items():
-        omega[key] = []
-        phi[key] = []
-        kappa[key] = []
-        for i, cam in enumerate(cam_list):
-            omega[key].append(cam.euler_angles[0])
-            phi[key].append(cam.euler_angles[1])
-            kappa[key].append(cam.euler_angles[2])
-
-    cam_ids = ["p1", "p2"]
-    fig, ax = plt.subplots(3, 2)
-    for i, cam_id in enumerate(cam_ids):
-        epoches = range(len(omega[cam_id]))
-        ax[0, i].plot(epoches, omega[cam_id] - omega[cam_id][0], "o")
-        ax[0, i].grid(visible=True, which="both")
-        ax[0, i].set_xlabel("Epoch")
-        ax[0, i].set_ylabel("Omega difference [deg]")
-
-        ax[1, i].plot(epoches, phi[cam_id] - phi[cam_id][0], "o")
-        ax[1, i].grid(visible=True, which="both")
-        ax[1, i].set_xlabel("Epoch")
-        ax[1, i].set_ylabel("Phi difference [deg]")
-
-        ax[2, i].plot(epoches, kappa[cam_id] - kappa[cam_id][0], "o")
-        ax[2, i].grid(visible=True, which="both")
-        ax[2, i].set_xlabel("Epoch")
-        ax[2, i].set_ylabel("Kappa difference [deg]")
-        fig.show()
 
 
 #%%
